Move Field into types (next to Container) to simplify dependency hell
